pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "beaver-bridge"
version = "0.1.0"
description = "Central API gateway bridging multiple MCP servers"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "mcp>=1.9",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "google-auth>=2.29",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
